+++ console/package.json
@@ -55,6 +55,7 @@
     "@types/styletron-standard": "^2.0.2",
     "@types/uuid": "^8.3.1",
     "@uiw/react-md-editor": "^3.9.3",
+    "@unocss/reset": "^0.54.1",
     "ahooks": "^3.7.8",
     "axios": "^0.21.1",
     "base64-js": "^1.5.1",
@@ -237,6 +238,7 @@
     "@typescript-eslint/eslint-plugin": "^5.23.0",
     "@typescript-eslint/parser": "^5.23.0",
     "@typescript-eslint/typescript-estree": "^5.23.0",
+    "@unocss/preset-typography": "^0.54.1",
     "@vitejs/plugin-react": "^2.0.0",
     "@welldone-software/why-did-you-render": "^7.0.1",
     "babel-eslint": "^10.1.0",
