+++ console/packages/starwhale-ui/src/TiptapEditor/extensions/index.tsx
@@ -1,7 +1,6 @@
 import StarterKit from '@tiptap/starter-kit'
 import HorizontalRule from '@tiptap/extension-horizontal-rule'
 import TiptapLink from '@tiptap/extension-link'
-import TiptapImage from '@tiptap/extension-image'
 import Placeholder from '@tiptap/extension-placeholder'
 import TiptapUnderline from '@tiptap/extension-underline'
 import TextStyle from '@tiptap/extension-text-style'
@@ -10,11 +9,12 @@ import TaskItem from '@tiptap/extension-task-item'
 import TaskList from '@tiptap/extension-task-list'
 import { Markdown } from 'tiptap-markdown'
 import Highlight from '@tiptap/extension-highlight'
-
-import SlashCommand from './slash-command'
 import { InputRule } from '@tiptap/core'
+import SlashCommand from './slash-command'
+
 // import UploadImagesPlugin from './plugins/upload-images'
 // import Paragraph from '@tiptap/extension-paragraph'
+// import TiptapImage from '@tiptap/extension-image'
 
 // const CustomImage = TiptapImage.extend({
 //     addProseMirrorPlugins() {
