+++ console/packages/starwhale-ui/src/TiptapEditor/index.tsx
@@ -3,52 +3,37 @@ import { useEditor, EditorContent } from '@tiptap/react'
 import { TiptapEditorProps } from './props'
 import { TiptapExtensions } from './extensions'
 import useLocalStorage from './hooks/use-local-storage'
-// import va from '@vercel/analytics'
 import DEFAULT_EDITOR_CONTENT from './default-content'
 import { EditorBubbleMenu } from './components'
-import { getPrevText } from './lib/editor'
-import { useDebounce } from 'react-use'
+import { useDebounceFn } from 'ahooks'
+import './styles/globals.css'
+import './styles/prosemirror.css'
 
 export default function TiptapEditor() {
     const [content, setContent] = useLocalStorage('content', DEFAULT_EDITOR_CONTENT)
     const [saveStatus, setSaveStatus] = useState('Saved')
-
     const [hydrated, setHydrated] = useState(false)
 
-    const [, debouncedUpdates] = useDebounce(async ({ editor }) => {
-        const json = editor.getJSON()
-        setSaveStatus('Saving...')
-        setContent(json)
-        // Simulate a delay in saving.
-        setTimeout(() => {
-            setSaveStatus('Saved')
-        }, 500)
-    }, 750)
+    const { run: debouncedUpdates } = useDebounceFn(
+        async ({ editor }) => {
+            const json = editor.getJSON()
+            setSaveStatus('Saving...')
+            // setContent(json)
+            setTimeout(() => {
+                setSaveStatus('Saved')
+            }, 500)
+        },
+        {
+            wait: 750,
+        }
+    )
 
     const editor = useEditor({
         extensions: TiptapExtensions,
         editorProps: TiptapEditorProps,
         onUpdate: (e) => {
             setSaveStatus('Unsaved')
-            const selection = e.editor.state.selection
-            const lastTwo = getPrevText(e.editor, {
-                chars: 2,
-            })
-            if (lastTwo === '++' && !isLoading) {
-                e.editor.commands.deleteRange({
-                    from: selection.from - 2,
-                    to: selection.from,
-                })
-                complete(
-                    getPrevText(e.editor, {
-                        chars: 5000,
-                    })
-                )
-                // complete(e.editor.storage.markdown.getMarkdown());
-                // va.track('Autocomplete Shortcut Used')
-            } else {
-                debouncedUpdates(e)
-            }
+            debouncedUpdates(e)
         },
         autofocus: 'end',
     })
